+++ extensions/cal-com-share-meeting-links/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Cal.com Share Meeting Links Changelog
 
-## [Initial Version] - 2022-06-28
+## [Initial Version] - 2022-08-05
+
+- Supports Cal.com cloud version
+- Copy meeting link to clipboard